AI-Accelerator-agents/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-client-cache==0.11.1
aiosignal==1.3.2
aiosqlite==0.20.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
async-lru==2.0.4
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.12.14
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
crashtest==0.4.1
cycler==0.12.1
deepdiff==8.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==7.2.1
dulwich==0.21.7
et_xmlfile==2.0.0
fastapi==0.115.6
fastjsonschema==2.21.1
feedparser==6.0.11
filelock==3.16.1
fonttools==4.55.3
frozendict==2.4.6
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-genai==0.6.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
groq==0.15.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==24.3.1
kiwisolver==1.4.8
lxml==5.3.0
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.23.0
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.2.1
openbb==4.3.5
openbb-benzinga==1.3.5
openbb-bls==1.0.3
openbb-cftc==1.0.3
openbb-commodity==1.2.6
openbb-core==1.3.8
openbb-crypto==1.3.5
openbb-currency==1.3.5
openbb-derivatives==1.3.5
openbb-econdb==1.2.5
openbb-economy==1.3.5
openbb-equity==1.3.5
openbb-etf==1.3.5
openbb-federal-reserve==1.3.5
openbb-fixedincome==1.3.5
openbb-fmp==1.3.5
openbb-fred==1.3.5
openbb-imf==1.0.2
openbb-index==1.3.5
openbb-intrinio==1.3.5
openbb-news==1.3.5
openbb-oecd==1.3.5
openbb-platform-api==1.0.4
openbb-polygon==1.3.5
openbb-regulators==1.3.5
openbb-sec==1.3.5
openbb-tiingo==1.3.5
openbb-tradingeconomics==1.3.5
openbb-us-eia==1.0.0
openbb-yfinance==1.3.6
openpyxl==3.1.5
orderly-set==5.2.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
peewee==3.17.8
pexpect==4.9.0
phidata==2.7.8
pillow==11.1.0
pkginfo==1.12.0
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
posthog==3.8.3
primp==0.10.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
ptyprocess==0.7.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.11.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.7.4
seaborn==0.13.2
setuptools==75.8.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
streamlit==1.41.1
tenacity==9.0.0
tldextract==5.1.3
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
trove-classifiers==2025.1.15.22
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.3.0
uuid7==0.1.0
uvicorn==0.32.1
virtualenv==20.29.0
watchdog==6.0.0
webencodings==0.5.1
websockets==14.1
xlrd==2.0.1
xmltodict==0.13.0
yarl==1.18.3
yfinance==0.2.51
youtube-transcript-api==0.6.3
zipp==3.21.0
